requirements.txt
streamlit
torch
torchvision
pillow
numpy
opencv-python
open_clip_torch